pyproject.toml
[tool.poetry]
name = "starkbank-backend-test"
version = "0.1.0"
description = ""
authors = ["João Pedro Prado <joaopedro0128@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
starkbank = "^2.22.0"
django = "^4.2.6"
boto3 = "^1.28.73"
celery = {extras = ["sqs"], version = "^5.3.4"}
faker = "^19.12.0"
debugpy = "^1.8.0"
djangorestframework = "^3.14.0"
pytest = "^7.4.3"
black = "^23.10.1"
ruff = "^0.1.3"
pytest-django = "^4.6.0"
celery-batches = "^0.8.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
gunicorn = "^21.2.0"
pre-commit = "^3.5.0"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["essential"], version = "^1.28.73"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=88
target-version=['py310']
extend-exclude='''
(
  ^/manage.py
  | ^/data/
  | ^/starkbank_backend/
  | migrations
)
'''

[tool.isort]
profile="black"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="starkbank_backend.settings"
python_files = ["test_*.py"]
